backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
psycopg[binary]>=3.2.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
email-validator>=2.2.0
python-multipart>=0.0.9
openai>=1.40.0
faiss-cpu>=1.8.0
numpy>=1.26.4
PyMuPDF>=1.24.9
reportlab>=4.2.2

